:recycle: Update middleware for hwWallet remove login moddleware and add hwWallet
