+++ src/store/middlewares/index.js
@@ -3,7 +3,7 @@ import peersMiddleware from './peers';
 import accountMiddleware from './account';
 import loadingBarMiddleware from './loadingBar';
 import offlineMiddleware from './offline';
-import loginMiddleware from './login';
+import hwWalletMiddleware from './hwWallet';
 // ToDo : enable this one when you solve the problem with multi account management
 // import notificationMiddleware from './notification';
 import votingMiddleware from './voting';
@@ -11,14 +11,14 @@ import followedAccountsMiddleware from './followedAccounts';
 import socketMiddleware from './socket';
 
 export default [
-  thunk,
-  peersMiddleware,
-  socketMiddleware,
+  // notificationMiddleware,
   accountMiddleware,
+  followedAccountsMiddleware,
+  hwWalletMiddleware,
   loadingBarMiddleware,
   offlineMiddleware,
-  loginMiddleware,
-  // notificationMiddleware,
+  peersMiddleware,
+  socketMiddleware,
+  thunk,
   votingMiddleware,
-  followedAccountsMiddleware,
 ];
